ref: fix sapling distribution
When sapling height distributions are predicted to strata which mean height is < 1.3m and mean diameter and mean age at breast height are 0.0, in some cases generated distributions can consist one or more trees whose
height is > 1.3 m. In those cases diameter is estimated with models but age at breast height is still 0. Simple solution for that problem is to set up age at breast height to 1 as, for example, in VMI12 data this only concerns generated saplings which maximum height was 1,56 m. This solution has been added to finalize_trees -function in tree_generation.py module.